Move IO-related code from the parsing module.
