+++ src/dlock/processing.py
@@ -20,8 +20,9 @@ from __future__ import annotations
 import dataclasses
 from typing import List, Optional
 
+from dlock.io import Dockerfile
 from dlock.output import Log
-from dlock.parsing import Dockerfile, FromInstruction, Instruction, parse_dockerfile
+from dlock.parsing import FromInstruction, Instruction, parse_dockerfile
 from dlock.registry import Resolver
